eliminar un cliente completada

+++ proyectoFinalPruebaAlcantarillaCaladoCandida/src/main/java/com/salesianostriana/dam/proyectofinalprueba/controller/AdminControlador.java
@@ -12,8 +12,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.salesianostriana.dam.proyectofinalprueba.model.Administrador;
+import com.salesianostriana.dam.proyectofinalprueba.model.Categoria;
 import com.salesianostriana.dam.proyectofinalprueba.model.Cliente;
 import com.salesianostriana.dam.proyectofinalprueba.model.Usuario;
+import com.salesianostriana.dam.proyectofinalprueba.service.CategoriaService;
 import com.salesianostriana.dam.proyectofinalprueba.service.ClienteService;
 
 @Controller
@@ -23,7 +25,7 @@ public class AdminControlador {
 	@Autowired
 	private ClienteService clienteServ;
 		
-	@GetMapping("/listaCliente")
+	@GetMapping("/listaCliente/")
 	public String mostrarClientes(Model model) {
 		model.addAttribute("listaCliente", clienteServ.findAll());
 		return "/admin/listaClientes";
@@ -33,38 +35,46 @@ public class AdminControlador {
 	public String agregarCliente(Model model) {
 		model.addAttribute("cliente", new Cliente());
 		model.addAttribute("editar", true);
+		model.addAttribute("password", "password");
 		return "/admin/formCliente";
 	}
 	
 	@PostMapping("/agregarCliente/submit")
 	public String submit(@ModelAttribute("cliente") Cliente cliente) {
-		clienteServ.encriptarContraseya(cliente);
-		clienteServ.save(cliente);
-		return "redirect:/admin/listaCliente";
+		clienteServ.save(cliente); 
+		return "redirect:/admin/listaCliente/";
 	}
 	
 	@GetMapping("/editarCliente/{id}")
 	public String editarProducto(@PathVariable("id") Long id, Model model) {
 	
 		if(clienteServ.findById(id).isPresent()) {
-			model.addAttribute("cliente", clienteServ.findById(id).orElseThrow());
+			model.addAttribute("cliente", clienteServ.buscarClientePorId(id));
 			model.addAttribute("editar", false);
+			model.addAttribute("password", "hidden");
 			return "/admin/formCliente"; 
 		}else {
-			return"redirect:/admin/listaCliente";
+			return"redirect:/admin/listaCliente/";
 		}
 	}
 	
 	@PostMapping("/editarCliente/submit")
 	public String procesarEditar(@ModelAttribute("cliente") Cliente cliente) {
 		clienteServ.edit(cliente);
-		return "redirect:/admin/listaCliente"; 
+		return "redirect:/admin/listaCliente/"; 
 	}
 	
 	@GetMapping("/eliminarCliente/{id}")
 	public String eliminar(@PathVariable("id") Long id) {
-		clienteServ.deleteById(id);
-		return "redirect:/admin/listaCliente";
+		Cliente cliente = clienteServ.buscarClientePorId(id);
+		if(cliente != null) {
+			if(clienteServ.numVentasCliente(cliente) == 0 && clienteServ.numAdopcionCliente(cliente) == 0) {
+				clienteServ.deleteById(id);
+			}else {
+				return "redirect:/admin/listaCliente/?error=true";
+			}
+		}
+		return "redirect:/admin/listaCliente/";
 	} 
 	
 	@GetMapping("/perfilAdmin")
